tutorials: update video link and rename UDF
